pyproject.toml
[project]
name = "atc-transcribe-test"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper",
    "ffmpeg-python",
    "numpy>=1.24.0",
    "scipy",
    "silero-vad>=4.0",
    "torch",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true